+++ notes/app.R
@@ -764,9 +764,7 @@ ui <- fluidPage(
 
 
         column(5, conditionalPanel(
-          condition = "input.stage == 'Design stage' ||
-                   (input.by_treatment &&
-                   input.stage != 'Real-time after enrollment completion')",
+          condition = "input.stage == 'Design stage' || input.by_treatment",
 
           selectInput(
             "k", "Treatments",
@@ -836,6 +834,12 @@ server <- function(input, output, session) {
     } else {
       hideTab(inputId = "results", target = "observed_data_panel")
     }
+  })
+
+
+  # whether to allow the user to specify the number of treatments
+  observeEvent(input$stage, {
+    shinyjs::toggleState("k", input$stage == "Design stage")
   })
 
 
@@ -948,7 +952,19 @@ server <- function(input, output, session) {
   })
 
 
-  k <- reactive(as.numeric(input$k))
+  k <- reactive({
+    if (!input$by_treatment) {
+      k = 1
+    } else if (input$stage != "Design stage" && !is.null(df())) {
+      k = length(table(df()$treatment))
+      updateSelectInput(session, "k", selected=k)
+    } else {
+      k = as.numeric(input$k)
+    }
+    k
+  })
+
+
 
 
   treatment_allocation <- reactive({
@@ -1339,7 +1355,7 @@ server <- function(input, output, session) {
   # event fit
   event_fit <- reactive({
     if (!is.null(df())) {
-      if (!input$by_treatment) {
+      if (!input$by_treatment || k() == 1) {
         event_fit <- fitEvent(
           df(), input$event_model, piecewiseSurvivalTime(),
           spline_k(), input$spline_scale, showplot = FALSE)
@@ -1359,7 +1375,7 @@ server <- function(input, output, session) {
   # dropout fit
   dropout_fit <- reactive({
     if (!is.null(df()) && input$dropout_model != "None") {
-      if (!input$by_treatment) {
+      if (!input$by_treatment || k() == 1) {
         shiny::validate(
           need(observed()$c0 > 0, paste("The number of dropouts must be",
                                         "positive to fit a dropout model.")))
@@ -1718,11 +1734,7 @@ server <- function(input, output, session) {
         need(!is.null(df()),
              "Please upload data for real-time prediction."))
 
-      if (input$by_treatment) {
-        k = length(unique(df()$treatment))
-      } else {
-        k = 1
-      }
+      k = k()
 
       if (!input$by_treatment || k == 1) {
         if (to_predict() == "Enrollment only") {
@@ -1820,10 +1832,6 @@ server <- function(input, output, session) {
             need(target_n() > observed()$n0,
                  "Target enrollment has been reached."))
 
-          shiny::validate(
-            need(k() == k,
-                 "Number of treatments must match observed."))
-
           alloc = treatment_allocation()
 
           pred1 <- getPrediction(
@@ -1895,26 +1903,25 @@ server <- function(input, output, session) {
             dplyr::slice(dplyr::n()) %>%
             dplyr::group_by(treatment)
 
-          # extend observed to cutoff date
-          dfa1 <- dfa %>%
-            dplyr::slice(dplyr::n()) %>%
-            dplyr::mutate(t = t0)
-
-          if (max(dfa$t) < t0) {
-            dfa <- dfa %>%
-              dplyr::bind_rows(dfa1)
-          }
-
-
-          # concatenate subjects enrolled before and after data cut
-          # add day 1
+          # add day 1 and extend observed to cutoff date
           df0 <- dplyr::tibble(treatment = 1:k,
                                t = 1, n = 0, lower = NA, upper = NA,
                                mean = 0, var = 0)
 
-          # concatenate subjects enrolled before and after data cut
-          dfs <- df0 %>%
+          dfa1 <- dfa %>%
+            dplyr::slice(dplyr::n()) %>%
+            dplyr::mutate(t = t0)
+
+          dfa <- df0 %>%
             dplyr::bind_rows(dfa) %>%
+            dplyr::bind_rows(dfa1) %>%
+            dplyr::group_by(treatment, t) %>%
+            dplyr::slice(dplyr::n()) %>%
+            dplyr::group_by(treatment)
+
+
+          # concatenate subjects enrolled before and after data cut
+          dfs <- dfa %>%
             dplyr::bind_rows(dfb) %>%
             dplyr::mutate(date = as.Date(t - 1, origin = trialsdt))
 
@@ -1993,7 +2000,7 @@ server <- function(input, output, session) {
 
           df <- df() %>%
             dplyr::mutate(arrivalTime = as.numeric(randdt - trialsdt + 1),
-                          totalTime = arrivalTime + time)
+                          totalTime = arrivalTime + time - 1)
 
           shiny::validate(
             need(target_n() > n0,
@@ -2017,9 +2024,6 @@ server <- function(input, output, session) {
                 "to fit a dropout model for each treatment.")))
           }
 
-          shiny::validate(
-            need(k() == k,
-                 "Number of treatments must match observed."))
 
           # predict enrollment
           enroll_fit <- fitEnrollment(
@@ -2090,24 +2094,25 @@ server <- function(input, output, session) {
             dplyr::slice(dplyr::n()) %>%
             dplyr::group_by(treatment)
 
-          # extend observed to cutoff date
-          dfa1 <- dfa %>%
-            dplyr::slice(dplyr::n()) %>%
-            dplyr::mutate(t = t0)
-
-          if (max(dfa$t) < t0) {
-            dfa <- dfa %>%
-              dplyr::bind_rows(dfa1)
-          }
-
-          # add day 1
+          # add day 1 and extend observed to cutoff date
           df0 <- dplyr::tibble(treatment = 1:k,
                                t = 1, n = 0, lower = NA, upper = NA,
                                mean = 0, var = 0)
 
-          # concatenate subjects enrolled before and after data cut
-          enroll_pred_df <- df0 %>%
+          dfa1 <- dfa %>%
+            dplyr::slice(dplyr::n()) %>%
+            dplyr::mutate(t = t0)
+
+          dfa <- df0 %>%
             dplyr::bind_rows(dfa) %>%
+            dplyr::bind_rows(dfa1) %>%
+            dplyr::group_by(treatment, t) %>%
+            dplyr::slice(dplyr::n()) %>%
+            dplyr::group_by(treatment)
+
+
+          # concatenate subjects enrolled before and after data cut
+          enroll_pred_df <- dfa %>%
             dplyr::bind_rows(dfb) %>%
             dplyr::mutate(date = as.Date(t - 1, origin = trialsdt)) %>%
             dplyr::mutate(parameter = "Enrollment") %>%
@@ -2259,6 +2264,10 @@ server <- function(input, output, session) {
                              mean = mean(natrisk),
                              var = var(natrisk))
 
+          # day 1
+          df0 <- dplyr::tibble(t = 1, n = 0, lower = NA, upper = NA,
+                               mean = 0, var = 0)
+
           # observed number of events before data cut
           dfa <- df %>%
             dplyr::arrange(totalTime) %>%
@@ -2271,6 +2280,12 @@ server <- function(input, output, session) {
             dplyr::group_by(t) %>%
             dplyr::slice(dplyr::n()) %>%
             dplyr::ungroup()
+
+          # add day 1
+          if (min(dfa$t) > 1) {
+            dfa <- df0 %>%
+              dplyr::bind_rows(dfa)
+          }
 
 
           # observed number of dropouts before data cut
@@ -2285,6 +2300,13 @@ server <- function(input, output, session) {
             dplyr::group_by(t) %>%
             dplyr::slice(dplyr::n()) %>%
             dplyr::ungroup()
+
+          # add day 1
+          if (min(dfc$t) > 1) {
+            dfc <- df0 %>%
+              dplyr::bind_rows(dfc)
+          }
+
 
           # observed number of subjects at risk before cutoff
           t2 = setdiff(sort(unique(c(df$arrivalTime, df$totalTime))), t0)
@@ -2301,27 +2323,27 @@ server <- function(input, output, session) {
                                            lower = NA, upper = NA,
                                            mean = r0, var = 0))
 
-          # day 1
-          df0 <- dplyr::tibble(t = 1, n = 0, lower = NA, upper = NA,
-                               mean = 0, var = 0)
+          # add day 1
+          if (min(dfe$t) > 1) {
+            dfe <- df0 %>%
+              dplyr::bind_rows(dfe)
+          }
+
 
           # add day 1 and concatenate events before and after data cut
-          event_pred_df1 <- df0 %>%
-            dplyr::bind_rows(dfa) %>%
+          event_pred_df1 <- dfa %>%
             dplyr::bind_rows(dfb) %>%
             dplyr::mutate(date = as.Date(t - 1, origin = trialsdt)) %>%
             dplyr::mutate(treatment = 9999, parameter = "Event")
 
           # add day 1 and concatenate dropouts before and after data cut
-          dropout_pred_df1 <- df0 %>%
-            dplyr::bind_rows(dfc) %>%
+          dropout_pred_df1 <- dfc %>%
             dplyr::bind_rows(dfd) %>%
             dplyr::mutate(date = as.Date(t - 1, origin = trialsdt)) %>%
             dplyr::mutate(treatment = 9999, parameter = "Dropout")
 
           # add day 1 and concatenate ongoing before and after data cut
-          ongoing_pred_df1 <- df0 %>%
-            dplyr::bind_rows(dfe) %>%
+          ongoing_pred_df1 <- dfe %>%
             dplyr::bind_rows(dff) %>%
             dplyr::mutate(date = as.Date(t - 1, origin = trialsdt)) %>%
             dplyr::mutate(treatment = 9999, parameter = "Ongoing")
@@ -2415,7 +2437,7 @@ server <- function(input, output, session) {
 
           df <- df() %>%
             dplyr::mutate(arrivalTime = as.numeric(randdt - trialsdt + 1),
-                          totalTime = arrivalTime + time)
+                          totalTime = arrivalTime + time - 1)
 
           # enrollment by treatment and overall
           dfpooled <- df %>% dplyr::mutate(treatment = 9999)
@@ -2432,15 +2454,22 @@ server <- function(input, output, session) {
             dplyr::slice(dplyr::n()) %>%
             dplyr::group_by(treatment)
 
-          # extend observed to cutoff date
+          # add day 1 and extend observed to cutoff date
+          df0 <- dplyr::tibble(treatment = c(1:k, 9999),
+                               t = 1, n = 0, lower = NA, upper = NA,
+                               mean = 0, var = 0)
+
           dfa1 <- dfa %>%
             dplyr::slice(dplyr::n()) %>%
             dplyr::mutate(t = t0)
 
-          if (max(dfa$t) < t0) {
-            dfa <- dfa %>%
-              dplyr::bind_rows(dfa1)
-          }
+          dfa <- df0 %>%
+            dplyr::bind_rows(dfa) %>%
+            dplyr::bind_rows(dfa1) %>%
+            dplyr::group_by(treatment, t) %>%
+            dplyr::slice(dplyr::n()) %>%
+            dplyr::group_by(treatment)
+
 
           # add predicted from data cut to specified years after data cut
           dfb1 <- dfa1 %>%
@@ -2449,13 +2478,8 @@ server <- function(input, output, session) {
           dfb2 <- dfb1 %>%
             dplyr::mutate(t = t0 + 365*nyears())
 
-          # add day 1
-          df0 <- dplyr::tibble(treatment = c(1:k, 9999),
-                               t = 1, n = 0, lower = NA, upper = NA,
-                               mean = 0, var = 0)
 
-          enroll_pred_df <- df0 %>%
-            dplyr::bind_rows(dfa) %>%
+          enroll_pred_df <- dfa %>%
             dplyr::bind_rows(dfb1) %>%
             dplyr::bind_rows(dfb2) %>%
             dplyr::mutate(date = as.Date(t - 1, origin = trialsdt)) %>%
@@ -2609,6 +2633,10 @@ server <- function(input, output, session) {
                              mean = mean(natrisk),
                              var = var(natrisk))
 
+          # day 1
+          df0 <- dplyr::tibble(t = 1, n = 0, lower = NA, upper = NA,
+                               mean = 0, var = 0)
+
           # observed number of events before data cut
           dfa <- df %>%
             dplyr::arrange(totalTime) %>%
@@ -2621,6 +2649,12 @@ server <- function(input, output, session) {
             dplyr::group_by(t) %>%
             dplyr::slice(dplyr::n()) %>%
             dplyr::ungroup()
+
+          # add day 1
+          if (min(dfa$t) > 1) {
+            dfa <- df0 %>%
+              dplyr::bind_rows(dfa)
+          }
 
 
           # observed number of dropouts before data cut
@@ -2635,6 +2669,13 @@ server <- function(input, output, session) {
             dplyr::group_by(t) %>%
             dplyr::slice(dplyr::n()) %>%
             dplyr::ungroup()
+
+          # add day 1
+          if (min(dfc$t) > 1) {
+            dfc <- df0 %>%
+              dplyr::bind_rows(dfc)
+          }
+
 
           # observed number of subjects at risk before cutoff
           t2 = setdiff(sort(unique(c(df$arrivalTime, df$totalTime))), t0)
@@ -2651,28 +2692,27 @@ server <- function(input, output, session) {
                                            lower = NA, upper = NA,
                                            mean = r0, var = 0))
 
-          # day 1
-          df0 <- dplyr::tibble(t = 1, n = 0, lower = NA, upper = NA,
-                               mean = 0, var = 0)
+          # add day 1
+          if (min(dfe$t) > 1) {
+            dfe <- df0 %>%
+              dplyr::bind_rows(dfe)
+          }
 
 
           # add day 1 and concatenate events before and after data cut
-          event_pred_df1 <- df0 %>%
-            dplyr::bind_rows(dfa) %>%
+          event_pred_df1 <- dfa %>%
             dplyr::bind_rows(dfb) %>%
             dplyr::mutate(date = as.Date(t - 1, origin = trialsdt)) %>%
             dplyr::mutate(treatment = 9999, parameter = "Event")
 
           # add day 1 and concatenate dropouts before and after data cut
-          dropout_pred_df1 <- df0 %>%
-            dplyr::bind_rows(dfc) %>%
+          dropout_pred_df1 <- dfc %>%
             dplyr::bind_rows(dfd) %>%
             dplyr::mutate(date = as.Date(t - 1, origin = trialsdt)) %>%
             dplyr::mutate(treatment = 9999, parameter = "Dropout")
 
           # add day 1 and concatenate ongoing before and after data cut
-          ongoing_pred_df1 <- df0 %>%
-            dplyr::bind_rows(dfe) %>%
+          ongoing_pred_df1 <- dfe %>%
             dplyr::bind_rows(dff) %>%
             dplyr::mutate(date = as.Date(t - 1, origin = trialsdt)) %>%
             dplyr::mutate(treatment = 9999, parameter = "Ongoing")
@@ -2754,7 +2794,7 @@ server <- function(input, output, session) {
   output$statistics <- renderPrint({
     if (!is.null(df())) {
 
-      if (input$by_treatment) {
+      if (input$by_treatment && k() > 1) {
         if (to_predict() == "Enrollment and event" ||
             to_predict() == "Event only") {
           sum_by_trt <- df() %>%
@@ -2826,7 +2866,7 @@ server <- function(input, output, session) {
   output$cum_accrual_plot <- renderPlotly({
     cum_accrual_plot <- observed()$cum_accrual_plot
     if (!is.null(cum_accrual_plot)) {
-      if (!input$by_treatment) {
+      if (!input$by_treatment || k() == 1) {
         cum_accrual_plot
       } else {
         trialsdt = observed()$trialsdt
@@ -2939,7 +2979,7 @@ server <- function(input, output, session) {
   output$event_km_plot <- renderPlotly({
     event_km_plot <- observed()$event_km_plot
     if (!is.null(event_km_plot)) {
-      if (!input$by_treatment) {
+      if (!input$by_treatment || k() == 1) {
         event_km_plot
       } else {
         adtte <- observed()$adtte
@@ -2947,13 +2987,16 @@ server <- function(input, output, session) {
                                             treatment, data = adtte)
         treatment <- attr(event_km_fit$strata, "names")
 
-        event_km_df <- dplyr:: tibble(
+        event_km_df <- dplyr::tibble(
           treatment = treatment,
-          time = 0, surv = 1) %>%
+          time = 1, surv = 1) %>%
           dplyr::bind_rows(dplyr::tibble(
             treatment = rep(treatment, event_km_fit$strata),
             time = event_km_fit$time,
-            surv = event_km_fit$surv))
+            surv = event_km_fit$surv)) %>%
+          dplyr::group_by(treatment, time) %>%
+          dplyr::slice(dplyr::n()) %>%
+          dplyr::ungroup()
 
         event_km_plot <- plotly::plot_ly(event_km_df, x=~time, y=~surv,
                                          linetype=~treatment) %>%
@@ -2976,7 +3019,7 @@ server <- function(input, output, session) {
   output$dropout_km_plot <- renderPlotly({
     dropout_km_plot <- observed()$dropout_km_plot
     if (!is.null(dropout_km_plot)) {
-      if (!input$by_treatment) {
+      if (!input$by_treatment || k() == 1) {
         dropout_km_plot
       } else {
         adtte <- observed()$adtte
@@ -2984,13 +3027,16 @@ server <- function(input, output, session) {
                                               treatment, data = adtte)
         treatment <- attr(dropout_km_fit$strata, "names")
 
-        dropout_km_df <- dplyr:: tibble(
+        dropout_km_df <- dplyr::tibble(
           treatment = treatment,
-          time = 0, surv = 1) %>%
+          time = 1, surv = 1) %>%
           dplyr::bind_rows(dplyr::tibble(
             treatment = rep(treatment, dropout_km_fit$strata),
             time = dropout_km_fit$time,
-            surv = dropout_km_fit$surv))
+            surv = dropout_km_fit$surv)) %>%
+          dplyr::group_by(treatment, time) %>%
+          dplyr::slice(dplyr::n()) %>%
+          dplyr::ungroup()
 
         dropout_km_plot <- plotly::plot_ly(dropout_km_df, x=~time, y=~surv,
                                            linetype=~treatment) %>%
@@ -3023,7 +3069,7 @@ server <- function(input, output, session) {
 
   output$event_fit1 <- renderPlotly({
     if (!is.null(event_fit())) {
-      if (!input$by_treatment) {
+      if (!input$by_treatment || k() == 1) {
         event_fit()$event_fit_plot
       } else {
         k = length(event_fit())
@@ -3054,7 +3100,7 @@ server <- function(input, output, session) {
 
   output$dropout_fit1 <- renderPlotly({
     if (!is.null(dropout_fit())) {
-      if (!input$by_treatment) {
+      if (!input$by_treatment || k() == 1) {
         dropout_fit()$dropout_fit_plot
       } else {
         k = length(dropout_fit())
@@ -3841,9 +3887,9 @@ server <- function(input, output, session) {
 
     load(file=file$datapath)
 
-    if (x$stage == 'Design stage' ||
-        (x$by_treatment &&
-         x$stage != 'Real-time after enrollment completion')) {
+    if ((x$stage == 'Design stage' ||
+         (x$by_treatment &&
+          x$stage != 'Real-time after enrollment completion')) && x$k > 1) {
       updateMatrixInput(
         session, paste0("treatment_allocation_", x$k),
         value=matrix(x$treatment_allocation, ncol = 1,
@@ -3982,7 +4028,7 @@ server <- function(input, output, session) {
                            "Starting time")))
         } else if (x$event_model == "Spline") {
           updateNumericInput(session, "spline_k", value=x$spline_k)
-          updateRadioButtons(session, "spline_scale", value=x$spline_scale)
+          updateRadioButtons(session, "spline_scale", selected=x$spline_scale)
         }
 
         updateRadioButtons(session, "dropout_model", selected=x$dropout_model)
@@ -4023,9 +4069,7 @@ server <- function(input, output, session) {
 
     updateCheckboxInput(session, "by_treatment", value=x$by_treatment)
 
-    if (x$stage == 'Design stage' ||
-        (x$by_treatment &&
-         x$stage != 'Real-time after enrollment completion')) {
+    if (x$stage == 'Design stage' || x$by_treatment) {
       updateSelectInput(session, "k", selected=x$k)
     }